requirements.txt
Flask==0.10.1
requests==1.2.3
celery==3.1.13
redis==2.6.2
hiredis==0.1.1
PyYAML==3.10
lxml==2.3.5
Flask-Cache==0.11
raven==4.0.3
blinker==1.2
Flask-StatsD==0.1.1
suds==0.4
requests-oauthlib==0.3.2
Flask-SQLAlchemy==1.0
rdflib==4.1.1
geojson==1.0.6
Shapely==1.3.0
